Recording statistics per test (identified with testTimeStamp) removed the concept of mode alias - modes with aliases are now treated as the same mode and named as Ionian/Major

+++ AncientModes/NSMutableArray+Shuffling.m
@@ -13,7 +13,7 @@
     NSUInteger count = [self count];
     for (uint i = 0; i < count; ++i){
         // Select a random element between i and end of array to swap with.
-        int nElements = count - i;
+        int nElements = (int)count - i;
         int n = arc4random_uniform(nElements) + i;
         [self exchangeObjectAtIndex:i withObjectAtIndex:n];
     }
